upload

+++ database/migrations/2019_11_13_050017_create_order_details_table.php
@@ -16,7 +16,7 @@ class CreateOrderDetailsTable extends Migration
         Schema::create('order_details', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->bigInteger('product_id')->unsigned()->nullable();
-            $table->tinyInteger('amount');
+            $table->integer('amount');
             $table->bigInteger('order_id')->unsigned();;
             $table->foreign('product_id')->references('id')->on('products')->onDelete('set null');
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
